Narrow projection-effects PR to standalone richness and lensing bias models
  This update responds to review feedback by removing the combined
  projection+mass-richness classes (CostanziBinned and CostanziUnBinned) from the PR.
  The projection model is now kept as a standalone richness-bias kernel, while the full integration over rich_tru, mass, redshift, purity, and observed-richness bins is left for a future recipe-level PR.

  Main changes relative to LSSTDESC/crow main:

  - Add crow.cluster_modules.projection_effects.
  - Add richness_bias.py with CostanziRichnessBias:
    - P(rich_obs | rich_tru)
    - observed-richness-bin integral at fixed rich_tru (this is useful for calculating the number counts over a observed richness bin?)
    - default Costanzi projection parameters (this can be removed in the future)
  - Add lensing_bias.py with CostanziLensingBias / bsel:
    - multiplicative post-stacking DeltaSigma correction
    - explicit radius handling: comoving Mpc/h or physical Mpc converted using z_ref and h
  - Wire optional lensing_profile_correction into binned recipes:
    - applied after stacked lensing integration
    - restricted to DeltaSigma
    - raises for reduced shear
  - Add projection tutorial notebook.
  - Add tests for:
    - Costanzi richness-bias probability shape, positivity, finiteness
    - observed-richness-bin integration at fixed rich_tru
    - parameter validation
    - lensing-bias radius conversion and shape validation
    - exact/grid recipe DeltaSigma correction behavior
    - reduced-shear rejection
  - Remove the earlier CostanziBinned / CostanziUnBinned classes.
  - Remove the temporary grid projection-cache integration from this PR.
